user 정보 로직 userContext로 이동, 페이지 새로고침 없이 user 다시 불러오도록 설정
